requirements.txt
python-telegram-bot==20.7
pillow==10.0.1
